refactor: type 분리 - UseSetting, API 부분 각 컴포넌트 Type 분리
