+++ api/profile.tsx
@@ -1,22 +1,8 @@
 import api from "./config";
-
-/* 타입 정의 */
-export interface CreateProfileRequest {
-  display_name: string;
-  gender: string;
-  birth_date: string;
-  bio_note: string;
-  age: number;
-  image: string;
-}
-
-export interface CreateProfileResponse {
-  status_code: number;
-  message: string;
-  option?: {
-    data?: any;
-  };
-}
+import type {
+  CreateProfileRequest,
+  CreateProfileResponse,
+} from "./types/profile";
 
 /* 프로필 API */
 // 프로필 생성
