+++ api/types/index.ts
@@ -1,0 +1,16 @@
+/* 모든 타입 정의 내보내기 */
+
+// 지역 관련 타입
+export * from "./area";
+
+// 인증 관련 타입
+export * from "./auth";
+
+// 프로필 관련 타입
+export * from "./profile";
+
+// 태그 관련 타입
+export * from "./tag";
+
+// 사용자 설정 관련 타입
+export * from "./userSetting";
