+++ components/useSetting/DistanceStep.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components/native";
 import Slider from "@react-native-community/slider";
-import { UserSettingData } from "./NameStep";
+import type { UserSettingData, DistanceStepProps } from "./types";
 import { unClickColor, secondaryColor, textColor } from "../../styles/Color";
 
 const Container = styled.View`
@@ -36,12 +36,6 @@ const ValueText = styled.Text`
   font-family: Paperlogy-SemiBold;
   color: ${secondaryColor};
 `;
-
-interface DistanceStepProps {
-  data: UserSettingData;
-  onDataChange: (data: Partial<UserSettingData>) => void;
-  onValidationChange: (isValid: boolean) => void;
-}
 
 export default function DistanceStep({
   data,
