+++ components/useSetting/GoalStep.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components/native";
 import BrandTextField from "../input/BrandTextField";
-import { UserSettingData } from "./NameStep";
+import type { UserSettingData, GoalStepProps } from "./types";
 
 const Container = styled.View`
   flex: 1;
@@ -20,12 +20,6 @@ const Question = styled.Text`
 const Answer = styled.View`
   align-self: stretch;
 `;
-
-interface GoalStepProps {
-  data: UserSettingData;
-  onDataChange: (data: Partial<UserSettingData>) => void;
-  onValidationChange: (isValid: boolean) => void;
-}
 
 export default function GoalStep({
   data,
