+++ components/useSetting/types/index.ts
@@ -1,0 +1,3 @@
+/* 모든 사용자 설정 타입 내보내기 */
+
+export * from "./userSetting";
